Lab 4: Fix incorrect calculation of string length
Using `sizeof(string) / sizeof(string[0])` is only possible when
`string` is an array that is directly accessible. When accessing an
array through a pointer, `sizeof(string)` will return the size of the
pointer instead of the size of the array, leading to the length always
being calculated as 4 in this case.

To get around this, we can just avoid using lengths entirely, since we
assume that the string passed to the function is null-terminated. We
just loop through each character using a pointer and stop once we hit
the null character.